fixed downloading and setting wallpaper on Android Q
