+++ app/src/main/java/com/droidheat/amoledbackgrounds/DownloadsFragment.java
@@ -1,12 +1,8 @@
 package com.droidheat.amoledbackgrounds;
 
 import android.Manifest;
-import android.content.Context;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
-import androidx.core.app.ActivityCompat;
-import androidx.fragment.app.Fragment;
-import androidx.core.content.ContextCompat;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -14,104 +10,61 @@ import android.widget.Button;
 import android.widget.GridView;
 import android.widget.Toast;
 
+import androidx.activity.result.ActivityResultLauncher;
+import androidx.activity.result.contract.ActivityResultContracts;
+import androidx.core.content.ContextCompat;
+import androidx.fragment.app.Fragment;
+
+import com.droidheat.amoledbackgrounds.adapters.MyDownloadsAdapter;
+
 public class DownloadsFragment extends Fragment {
 
-    public DownloadsFragment() {
+  public DownloadsFragment() {
+  }
+
+  MyDownloadsAdapter myDownloadsAdapter;
+  GridView gridView;
+  Button permissionBtn;
+
+  @Override
+  public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                           Bundle savedInstanceState) {
+    View view = inflater.inflate(R.layout.fragment_downloads, container, false);
+
+    gridView = view.findViewById(R.id.gridView);
+    permissionBtn = view.findViewById(R.id.button_permission);
+
+    myDownloadsAdapter = new MyDownloadsAdapter(getActivity());
+    gridView.setAdapter(myDownloadsAdapter);
+
+    final String PERMISSION = Manifest.permission.READ_EXTERNAL_STORAGE;
+
+    if (ContextCompat.checkSelfPermission(requireActivity(), PERMISSION) != PackageManager.PERMISSION_GRANTED) {
+      gridView.setVisibility(View.INVISIBLE);
+      requestPermissionLauncher.launch(PERMISSION);
+    } else {
+      permissionBtn.setVisibility(View.INVISIBLE);
+      gridView.setVisibility(View.VISIBLE);
     }
 
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
+    permissionBtn.setOnClickListener(v -> requestPermissionLauncher.launch(PERMISSION));
 
-    }
+    return view;
+  }
 
-    MyDownloadsAdapter myDownloadsAdapter;
-    GridView gridView;
-    final private int MY_PERMISSIONS_REQUEST = 621;
-    Button permissionBtn;
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_downloads, container, false);
-
-        gridView = view.findViewById(R.id.gridView);
-        permissionBtn = view.findViewById(R.id.button_permission);
-
-        myDownloadsAdapter = new MyDownloadsAdapter(getActivity());
-        gridView.setAdapter(myDownloadsAdapter);
-
-        permissionBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (ContextCompat.checkSelfPermission(getActivity(), Manifest.permission.WRITE_EXTERNAL_STORAGE)
-                        != PackageManager.PERMISSION_GRANTED && ContextCompat.checkSelfPermission(getActivity(), Manifest.permission.READ_EXTERNAL_STORAGE)
-                        != PackageManager.PERMISSION_GRANTED) {
-                    requestPermissions(
-                            new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE,Manifest.permission.READ_EXTERNAL_STORAGE},
-                            MY_PERMISSIONS_REQUEST);
-                } else {
-                    permissionBtn.setVisibility(View.INVISIBLE);
-                }
+  private final ActivityResultLauncher<String> requestPermissionLauncher =
+          registerForActivityResult(new ActivityResultContracts.RequestPermission(), isGranted -> {
+            if (isGranted) {
+              permissionBtn.setVisibility(View.INVISIBLE);
+              gridView.setVisibility(View.VISIBLE);
+            } else {
+              Toast.makeText(getActivity(), "Read Permission Denied", Toast.LENGTH_SHORT).show();
+              // Explain to the user that the feature is unavailable because the
+              // features requires a permission that the user has denied. At the
+              // same time, respect the user's decision. Don't link to system
+              // settings in an effort to convince the user to change their
+              // decision.
             }
-        });
+          });
 
-        if (ContextCompat.checkSelfPermission(getActivity(), Manifest.permission.WRITE_EXTERNAL_STORAGE)
-                != PackageManager.PERMISSION_GRANTED && ContextCompat.checkSelfPermission(getActivity(), Manifest.permission.READ_EXTERNAL_STORAGE)
-                != PackageManager.PERMISSION_GRANTED) {
-            requestPermissions(
-                    new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE,Manifest.permission.READ_EXTERNAL_STORAGE},
-                    MY_PERMISSIONS_REQUEST);
-        } else {
-            permissionBtn.setVisibility(View.INVISIBLE);
-        }
-
-        return view;
-    }
-
-    @Override
-    public void onRequestPermissionsResult(int requestCode,
-                                           String[] permissions, int[] grantResults) {
-        switch (requestCode) {
-            case MY_PERMISSIONS_REQUEST: {
-                // If request is cancelled, the result arrays are empty.
-                if (grantResults.length > 0
-                        && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-                    // permission was granted, yay! Do the
-                    // contacts-related task you need to do.
-                    myDownloadsAdapter.refresh();
-                    permissionBtn.setVisibility(View.INVISIBLE);
-                } else {
-                    // permission denied, boo! Disable the
-                    // functionality that depends on this permission.
-                    Toast.makeText(getContext(),"Permission required to use downloads tab", Toast.LENGTH_LONG).show();
-                }
-            }
-
-            // other 'case' lines to check for other
-            // permissions this app might request.
-        }
-    }
-
-    @Override
-    public void setUserVisibleHint(boolean isVisibleToUser) {
-        super.setUserVisibleHint(isVisibleToUser);
-        if (isVisibleToUser) {
-            try {
-                myDownloadsAdapter.refresh();
-            } catch (Exception ignored) {
-
-            }
-        }
-    }
-
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-    }
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-    }
 }
